# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/service/service_add_http": {
            "post": {
                "description": "添加HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "添加HTTP服务",
                "operationId": "/service/service_add_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAddHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_delete": {
            "get": {
                "description": "服务删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务删除",
                "operationId": "/service/service_delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_detail": {
            "get": {
                "description": "获取单个服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "获取单个服务",
                "operationId": "/service/service_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.ServiceDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_list": {
            "get": {
                "description": "服务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务列表",
                "operationId": "/service/service_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页的条数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServiceListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_stat": {
            "get": {
                "description": "服务流量统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务流量统计",
                "operationId": "/service/service_stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServiceStatOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_update_http": {
            "post": {
                "description": "修改HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "修改HTTP服务",
                "operationId": "/service/service_update_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceUpdateHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dao.AccessControl": {
            "type": "object",
            "properties": {
                "black_list": {
                    "description": "黑名单id",
                    "type": "string"
                },
                "client_ip_flow_limit": {
                    "description": "客户端ip限流",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "open_auth": {
                    "description": "是否开启权限 1=开启",
                    "type": "integer"
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer"
                },
                "service_id": {
                    "description": "服务的id",
                    "type": "integer"
                },
                "white_host_name": {
                    "description": "主机白名单",
                    "type": "string"
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string"
                }
            }
        },
        "dao.GrpcRule": {
            "type": "object",
            "properties": {
                "header_transfor": {
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
        "dao.HttpRule": {
            "type": "object",
            "properties": {
                "header_transfor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "need_https": {
                    "type": "integer"
                },
                "need_strip_uri": {
                    "type": "integer"
                },
                "need_websocket": {
                    "type": "integer"
                },
                "rule": {
                    "type": "string"
                },
                "rule_type": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "url_rewrite": {
                    "type": "string"
                }
            }
        },
...
        "dao.ServiceDetail": {
            "type": "object",
            "properties": {
                "access_control": {
                    "type": "object",
                    "$ref": "#/definitions/dao.AccessControl"
                },
                "grpc": {
                    "type": "object",
                    "$ref": "#/definitions/dao.GrpcRule"
                },
                "http": {
                    "type": "object",
                    "$ref": "#/definitions/dao.HttpRule"
                },
                "info": {
                    "description": "基本信息",
                    "type": "object",
                    "$ref": "#/definitions/dao.ServiceInfo"
                },
                "load_balance": {
                    "type": "object",
                    "$ref": "#/definitions/dao.LoadBalance"
                },
                "tcp": {
                    "type": "object",
                    "$ref": "#/definitions/dao.TcpRule"
                }
            }
        },
        "dao.ServiceInfo": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "load_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ServiceAddHTTPInput": {
            "type": "object",
            "required": [
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer"
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string"
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer"
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer"
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer"
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer"
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer"
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string"
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer"
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string"
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer"
                },
                "service_name": {
                    "description": "服务名",
                    "type": "string"
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer"
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer"
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer"
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string"
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string"
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string"
                }
            }
        },
...
        "dto.ServiceListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceListItemOutput"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer",
                    "example": 56
                }
            }
        },
        "dto.ServiceStatOutput": {
            "type": "object",
            "properties": {
                "today": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        100
                    ]
                },
                "yesterday": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        100
                    ]
                }
            }
        },
        "dto.ServiceUpdateHTTPInput": {
            "type": "object",
            "required": [
                "id",
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer"
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string"
                },
                "id": {
                    "description": "服务id",
                    "type": "integer"
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string",
                    "example": "127.0.0.1:80"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer"
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer"
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer"
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer"
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer"
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string"
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer"
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string"
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer"
                },
                "service_name": {
                    "description": "服务名",
                    "type": "string"
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer"
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer"
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer"
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string"
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string"
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}